+++ src/spatialdata/models/__init__.py
@@ -6,10 +6,12 @@ from spatialdata.models._utils import (
     Z,
     force_2d,
     get_axes_names,
+    get_channel_names,
     get_channels,
     get_spatial_axes,
     points_dask_dataframe_to_geopandas,
     points_geopandas_to_dask_dataframe,
+    set_channel_names,
     validate_axes,
     validate_axis_name,
 )
@@ -49,6 +51,8 @@ __all__ = [
     "check_target_region_column_symmetry",
     "get_table_keys",
     "get_channels",
+    "get_channel_names",
+    "set_channel_names",
     "force_2d",
     "RasterSchema",
 ]